fixup for modules.order containing *.o's
As of: v6.1-rc6-24-gf65a486821cf
virtme-run --kdir $builddir is broken.
The cause is:

commit f65a486821cfd363833079b2a7b0769250ee21c9 (HEAD)
Author: Masahiro Yamada <masahiroy@kernel.org>
Date:   Sun Dec 11 22:04:07 2022 +0900

    kbuild: change module.order to list *.o instead of *.ko

    scripts/Makefile.build replaces the suffix .o with .ko, then
    scripts/Makefile.modpost calls the sed command to change .ko back
    to the original .o suffix.

    Instead of converting the suffixes back-and-forth, store the .o paths
    in modules.order, and replace it with .ko in 'make modules_install'.

    This avoids the unneeded sed command.

    Signed-off-by: Masahiro Yamada <masahiroy@kernel.org>
    Reviewed-by: Luis Chamberlain <mcgrof@kernel.org>

virtme-prep-kdir-mod is expecting *.ko's, and is now getting *.o's.
It does the copying, but the *.o's arent loadable.

Fix with another shell-var and a parameter substitution on it.  This
will also work for kernels before the cited patch, which already have
the ko's

Signed-off-by: Jim Cromie <jim.cromie@gmail.com>
--
v1 - used sed, now uses bash parameter substitution wo forking.